1차_통합본_v1.2/backend/requirements.txt
# FastAPI 및 서버
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 데이터베이스 (SQLite는 Python 기본 내장)
sqlalchemy==2.0.23

# MariaDB/MySQL 사용 시에만 필요 (선택사항)
# pymysql==1.1.0
# cryptography==41.0.7 !

# JWT 토큰
python-jose[cryptography]==3.3.0 # !
python-multipart==0.0.6 # !

# 비밀번호 해싱
bcrypt==4.1.1 # !

# 데이터 검증
pydantic==2.5.0
pydantic[email]==2.5.0 # !

# 환경 변수
python-dotenv==1.0.0

# HTTP 요청 (API 크롤링용)
requests==2.31.0

# 웹 크롤링 (백업용, 선택사항)
# selenium==4.15.2
# webdriver-manager==4.0.1
# beautifulsoup4==4.12.2
# lxml==4.9.3

# ===== AI 모델 (새로 추가) =====
# Hugging Face Transformers (모델 로딩)
transformers>=4.35.0

# PyTorch (딥러닝 프레임워크)
# CPU 버전 (기본)
torch>=2.1.0
torchvision>=0.16.0

# GPU 버전 사용 시 아래 주석 해제하고 위 torch 주석 처리
# torch>=2.1.0+cu118 --extra-index-url https://download.pytorch.org/whl/cu118
# torchvision>=0.16.0+cu118 --extra-index-url https://download.pytorch.org/whl/cu118

# SafeTensors (모델 파일 포맷)
safetensors>=0.4.0

# Accelerate (모델 로딩 최적화, 권장)
accelerate>=0.24.0

# SentencePiece (토크나이저, 일부 모델에서 필요)
sentencepiece>=0.1.99

# Tokenizers (빠른 토크나이저)
tokenizers>=0.15.0

# NumPy (배열 연산)
numpy>=1.24.0

# SciPy (과학 계산, 선택)
scipy>=1.11.0

# Gemini API
google-generativeai

# AI/ML - Random Forest
scikit-learn==1.3.2
joblib==1.3.2